requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pydantic-settings==2.2.1

sqlalchemy==2.0.30

flask==3.0.3
flask_sqlalchemy==3.1.1
flask_jwt_extended==4.6.0

flask_bcrypt==1.0.1
flask_wtf==1.2.1
flask_migrate==4.0.7

pytz==2024.1
gunicorn==22.0.0
